break up tests for configurator
